+++ index.php
@@ -14,7 +14,7 @@ function tcp_send_security_headers() {
 		header('X-Frame-Options: DENY');
 		header('Referrer-Policy: strict-origin-when-cross-origin');
 		header('Permissions-Policy: camera=(), microphone=(), geolocation=()');
-		header("Content-Security-Policy: default-src 'self' https:; object-src 'none'; base-uri 'self'; form-action 'self'; frame-ancestors 'none'; script-src 'self' 'unsafe-inline' https:; style-src 'self' 'unsafe-inline' https:; img-src 'self' https: data:; frame-src https:; connect-src 'self'");
+		header("Content-Security-Policy: default-src 'self'; object-src 'none'; base-uri 'self'; form-action 'self'; frame-ancestors 'none'; script-src 'self'; style-src 'self'; img-src 'self' data:; connect-src 'self'");
 	}
 }
 
@@ -30,32 +30,28 @@ tcp_send_security_headers();
 
 $company = tcp_get('c');
 $city = tcp_get('l');
-$title = tcp_get('t');
 $company_html = tcp_html($company);
 $city_html = tcp_html($city);
 $share_url_html = tcp_html(tcp_share_url($company, $city));
 ?>
-<html xmlns:og="http://ogp.me/ns#">
+<!doctype html>
+<html lang="en" xmlns:og="http://ogp.me/ns#">
 <head>
+	<meta charset="UTF-8" />
+	<meta name="viewport" content="width=device-width, initial-scale=1" />
    <title>TechCompanyPay</title>
-	<!-- SET FACEBOOK META -->
 	<meta property="og:title" content="TechCompanyPay"/>
     <meta property="og:url" content="<?php echo $share_url_html; ?>"/>
-    <meta property="og:image" content="https://d2dixsj1sor9iw.cloudfront.net/images/og.png"/>
     <meta property="og:site_name" content="TechCompanyPay"/>
     <meta property="og:description" content="A hack showing the titles and average pay of most of the tech giants employees with LinkedIn profiles."/>
 	
-	<!-- IMPORT CSS -->
-	<link href='https://fonts.googleapis.com/css?family=Droid+Sans' rel='stylesheet' type='text/css'>
-	<link href='https://d2dixsj1sor9iw.cloudfront.net/css/stylesheet.css' rel='stylesheet' type='text/css'>
-	<!-- IMPORT JAVASCRIPT -->
-	<script type='text/javascript' src='https://dorkzvyk9srha.cloudfront.net/js/jquery-1.6.2.min.js'></script>
-		<script type='text/javascript' src='https://d2dixsj1sor9iw.cloudfront.net/javascript/core2.js'></script>
+	<link href="assets/app.css" rel="stylesheet" />
+	<script src="assets/app.js" defer></script>
 </head>	
 <body>
 	<div id="header">
 	<div class="column">
-	<a href="/" id="logo"></a>
+	<a href="/" id="logo">TechCompanyPay</a>
 	</div>
 	</div>
 
@@ -65,24 +61,7 @@ $share_url_html = tcp_html(tcp_share_url($company, $city));
 the data 
 on the site. See <a href='#disclaimer'>the disclaimer</a> for more of this stuff.</p>
 
-	<button class='button' id='topcompany'>Top Company Info</button>
-	<div id='companywrap' style='display:none'>	
-	<div id='companydata'></div>
-	</div>
-	<script type='text/javascript'>
-	$("#topcompany").click(function(){
-		$("#companywrap").toggle();
-		var on = 1;
-		if ($("#topcompany").html() == 'Top Company Info'){
-			$("#topcompany").html('Top Company Info - Hide');
-		} else{
-			$("#topcompany").html('Top Company Info');
-		}
-	});
-	</script>
-	
-	
-	<form id="searchform" method="post" action=''> 
+		<form id="searchform" method="post" action="find.php">
 
 		<h2>Basic Search:</h2> 
 	<p> Enter a company and city and hit search.</p>
@@ -100,7 +79,7 @@ on the site. See <a href='#disclaimer'>the disclaimer</a> for more of this stuff
 			<input class='input' type="text" name="city" id="city" value="<?php echo $city_html; ?>" />
 			</td>
 			<td>
-				<input type='submit' style='margin-top:20px' id='submitbsearch' class='button'/>
+				<button type="submit" id="submitbsearch" class="button">Search salaries</button>
 				
 			</td>
 		</tr>
@@ -108,9 +87,9 @@ on the site. See <a href='#disclaimer'>the disclaimer</a> for more of this stuff
 	 
 	</div> 
 	</form>
-	    <div id="search_results"></div>
-		
-		<div style='clear:both;content: ".";display: block; margin-top:10px;'></div>
+		    <div id="search_results" role="status" aria-live="polite"></div>
+
+			<div class="clear"></div>
 	<div id='disclaimer'>
 		<p>This is a legal disclaimer, it's needed these days. The data on the site may not reflect the salary of actual people 
 working at these companies. For example if you click on a link to LinkedIn profiles of possible employees they may not be actually earning the specific amount. Most roles and titles 
@@ -118,44 +97,10 @@ are averaged to provide a guide. All data on this site was obtained via open sou
 information provided on this 
 site.</p>
 
-<p>Created by <a href='https://plus.google.com/112601017731183943047/posts'>Gareth Paul Jones</a>.</p>
+<p>Created by <a href="https://github.com/garethpaul">Gareth Paul Jones</a>.</p>
 
 	</div>
 	</div>
 
-	<div id='likes'>
-		
-			
-			<div id="fb-root"></div>
-			<script>(function(d, s, id) {
-			  var js, fjs = d.getElementsByTagName(s)[0];
-			  if (d.getElementById(id)) {return;}
-			  js = d.createElement(s); js.id = id;
-			  js.src = "https://connect.facebook.net/en_US/all.js#xfbml=1";
-			  fjs.parentNode.insertBefore(js, fjs);
-			}(document, 'script', 'facebook-jssdk'));
-			</script>
-			<div class="fb-like" data-href="<?php echo $share_url_html; ?>" data-send="true" data-width="450" data-show-faces="false"></div>
-			
-			<script type="text/javascript" src="https://apis.google.com/js/plusone.js"></script>
-			<g:plusone></g:plusone>
-
-	</div>
-	
-	<div id='message' class='column' style='display:none;'></div>
-	<script type='text/javascript'>
-	<?php
-	if ($company !== ''){
-		echo "ajax_search();";
-	}
-	?>
-	</script>
-
-<script type="text/javascript">
-var sc_project=7264690; 
-var sc_invisible=1; 
-var sc_security="94fdf985"; 
-</script>
-<script type="text/javascript" src="https://www.statcounter.com/counter/counter_xhtml.js"></script>
 </body>
 </html>